day 5 passing after big refactor

+++ 05/include/digest.h
@@ -2,15 +2,15 @@
 #include <stdlib.h>
 #include <stdio.h>
 
-int* ingest(int* size, FILE* fd) {
-  int* list = calloc(OP_LIST_SIZE, sizeof(int));
+long* ingest(int* size, FILE* fd) {
+  long* list = calloc(OP_LIST_SIZE, sizeof(long));
   int capacity = OP_LIST_SIZE;
   *size = 0;
   int i;
   while(fscanf(fd, "%d,?", &i) != EOF) {
     if (*size == capacity) {
       capacity *= 2;
-      list = realloc(list, capacity * sizeof(int));
+      list = realloc(list, capacity * sizeof(long));
     }
 
     list[(*size)++] = i;
